# back-office/openapi.json
{"openapi":"3.1.0","info":{"description":"FinalStake back-end","title":"FinalStake","version":"0.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/auth/admin/signin":{"post":{"operationId":"authenticateUserAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignInDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseString"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"tags":["Auth","Admin"]}},"/api/auth/confirmEmail/send":{"post":{"operationId":"confirmEmailSend","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEmailDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"tags":["Auth"]}},"/api/auth/confirmEmail/{token}":{"post":{"operationId":"confirmEmail","parameters":[{"in":"path","name":"token","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"tags":["Auth"]}},"/api/auth/resetPassword/send":{"post":{"operationId":"resetPasswordSend","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEmailDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"tags":["Auth"]}},"/api/auth/resetPassword/{token}":{"post":{"operationId":"resetPassword","parameters":[{"in":"path","name":"token","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserNewPasswordDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"tags":["Auth"]}},"/api/auth/signin":{"post":{"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignInDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseString"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"tags":["Auth"]}},"/api/auth/signup":{"post":{"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignUpDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"tags":["Auth"]}},"/api/role/all":{"get":{"operationId":"getAllRoles","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseListString"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["Admin"]}},"/api/user/admin/profile/{id}":{"get":{"operationId":"getProfileAdmin","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseUserViewAdminDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User","Admin"]}},"/api/user/admin/search":{"get":{"operationId":"searchAdmin_1","parameters":[{"description":"Zero-based page index (0..N)","in":"query","name":"page","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"description":"The size of the page to be returned","in":"query","name":"size","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","in":"query","name":"sort","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"in":"query","name":"username","required":false,"schema":{"type":"string","maxLength":40,"minLength":0}},{"in":"query","name":"email","required":false,"schema":{"type":"string","maxLength":255,"minLength":0}},{"in":"query","name":"id","required":false,"schema":{"type":"string","maxLength":36,"minLength":0}},{"in":"query","name":"enabled","required":false,"schema":{"type":"boolean"}},{"in":"query","name":"locked","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseUserViewAdminDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User","Admin"]}},"/api/user/admin/{id}":{"delete":{"operationId":"deleteProfile","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User","Admin"]},"put":{"operationId":"updateUserAdmin","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateAdminDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User","Admin"]}},"/api/user/profile":{"delete":{"operationId":"deleteProfileCurrent","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User"]},"get":{"operationId":"getProfileCurrent","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseUserViewDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User"]},"put":{"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User"]}},"/api/user/profile/{id}":{"get":{"operationId":"getProfile","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseUserViewDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User"]}},"/api/user/search":{"get":{"operationId":"search","parameters":[{"in":"query","name":"searchDTO","required":true,"schema":{"$ref":"#/components/schemas/UserSearchDTO"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseListUserListViewDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["User"]}},"/api/wallet/my":{"get":{"operationId":"getMyWallet","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseWalletViewDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["Wallet"]}},"/api/withdrawal":{"post":{"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalCreateDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseWithdrawalViewDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["Withdrawal"]}},"/api/withdrawal/admin/search":{"get":{"operationId":"searchAdmin","parameters":[{"in":"query","name":"status","required":false,"schema":{"type":"string","enum":["PENDING","APPROVED","REJECTED","PAID"]}},{"in":"query","name":"username","required":false,"schema":{"type":"string"}},{"in":"query","name":"pageable","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseWithdrawalViewDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["Withdrawal","Admin"]}},"/api/withdrawal/admin/{id}":{"get":{"operationId":"getOneAdmin","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseWithdrawalViewDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["Withdrawal","Admin"]},"put":{"operationId":"updateStatus","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalAdminUpdateDTO"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["Withdrawal","Admin"]}},"/api/withdrawal/my":{"get":{"operationId":"getMyWithdrawals","parameters":[{"description":"Zero-based page index (0..N)","in":"query","name":"page","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"description":"The size of the page to be returned","in":"query","name":"size","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","in":"query","name":"sort","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseWithdrawalViewDTO"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}},"description":"Internal Server Error"}},"security":[{"Authentication":[]}],"tags":["Withdrawal"]}}},"components":{"schemas":{"BaseResponseListString":{"type":"object","properties":{"data":{"type":"array","items":{"type":"string"}},"error":{"type":"string"},"success":{"type":"boolean"}}},"BaseResponseListUserListViewDTO":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserListViewDTO"}},"error":{"type":"string"},"success":{"type":"boolean"}}},"BaseResponseString":{"type":"object","properties":{"data":{"type":"string"},"error":{"type":"string"},"success":{"type":"boolean"}}},"BaseResponseUserViewAdminDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/UserViewAdminDTO"},"error":{"type":"string"},"success":{"type":"boolean"}}},"BaseResponseUserViewDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/UserViewDTO"},"error":{"type":"string"},"success":{"type":"boolean"}}},"BaseResponseVoid":{"type":"object","properties":{"data":{},"error":{"type":"string"},"success":{"type":"boolean"}}},"BaseResponseWalletViewDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/WalletViewDTO"},"error":{"type":"string"},"success":{"type":"boolean"}}},"BaseResponseWithdrawalViewDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/WithdrawalViewDTO"},"error":{"type":"string"},"success":{"type":"boolean"}}},"PageResponseUserViewAdminDTO":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserViewAdminDTO"}},"error":{"type":"string"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"success":{"type":"boolean"},"totalElements":{"type":"integer","format":"int64"}}},"PageResponseWithdrawalViewDTO":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/WithdrawalViewDTO"}},"error":{"type":"string"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"success":{"type":"boolean"},"totalElements":{"type":"integer","format":"int64"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"UserEmailDTO":{"type":"object","properties":{"email":{"type":"string","maxLength":255,"minLength":0}}},"UserListViewDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid","minLength":1},"username":{"type":"string","maxLength":40,"minLength":0}},"required":["id","username"]},"UserNewPasswordDTO":{"type":"object","properties":{"newPassword":{"type":"string","maxLength":40,"minLength":0}},"required":["newPassword"]},"UserSearchDTO":{"type":"object","properties":{"limit":{"type":"integer","format":"int32"},"username":{"type":"string"}},"required":["limit","username"]},"UserSignInDTO":{"type":"object","properties":{"password":{"type":"string","maxLength":40,"minLength":0},"username":{"type":"string","maxLength":40,"minLength":0}},"required":["password","username"]},"UserSignUpDTO":{"type":"object","properties":{"email":{"type":"string","maxLength":255,"minLength":0},"password":{"type":"string","maxLength":40,"minLength":0},"username":{"type":"string","maxLength":40,"minLength":0}},"required":["email","password","username"]},"UserUpdateAdminDTO":{"type":"object","properties":{"enabled":{"type":"boolean"},"locked":{"type":"boolean"},"roles":{"type":"array","items":{"type":"string"}}},"required":["enabled","locked","roles"]},"UserUpdateDTO":{"type":"object","properties":{"email":{"type":"string","maxLength":255,"minLength":0},"username":{"type":"string","maxLength":40,"minLength":0}},"required":["email","username"]},"UserViewAdminDTO":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time","minLength":1},"email":{"type":"string","maxLength":255,"minLength":0},"enabled":{"type":"boolean"},"id":{"type":"string","format":"uuid","minLength":1},"locked":{"type":"boolean"},"roles":{"type":"array","items":{"type":"string"}},"username":{"type":"string","maxLength":40,"minLength":0}},"required":["createdAt","email","enabled","id","locked","roles","username"]},"UserViewDTO":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time","minLength":1},"email":{"type":"string","maxLength":255,"minLength":0},"id":{"type":"string","format":"uuid","minLength":1},"roles":{"type":"array","items":{"type":"string"}},"username":{"type":"string","maxLength":40,"minLength":0}},"required":["createdAt","email","id","roles","username"]},"WalletViewDTO":{"type":"object","properties":{"balance":{"type":"number"},"currency":{"type":"string"}}},"WithdrawalAdminUpdateDTO":{"type":"object","properties":{"adminNote":{"type":"string"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED","PAID"]}},"required":["status"]},"WithdrawalCreateDTO":{"type":"object","properties":{"amount":{"type":"number"},"payoutDetails":{"type":"string","minLength":1},"payoutMethod":{"type":"string","minLength":1}},"required":["amount","payoutDetails","payoutMethod"]},"WithdrawalViewDTO":{"type":"object","properties":{"adminNote":{"type":"string"},"amount":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"id":{"type":"string","format":"uuid"},"payoutDetails":{"type":"string"},"payoutMethod":{"type":"string"},"processedAt":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED","PAID"]},"username":{"type":"string"}}}},"securitySchemes":{"Authentication":{"bearerFormat":"JWT","scheme":"bearer","type":"http"}}}}