# back-office/src/app/openapi/model/baseResponseWalletViewDTO.ts
/**
 * FinalStake
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WalletViewDTO } from './walletViewDTO';


export interface BaseResponseWalletViewDTO { 
    data?: WalletViewDTO;
    error?: string;
    success?: boolean;
}